config with window size and interval

+++ client/config/config.go
@@ -14,7 +14,7 @@ type config struct {
 	ConNum     int    `json:"con_num" flag:"con_num" usage:"" default:"10"`
 	SendWin    int    `json:"send_win" flag:"send_win" default:"1024" usage:""`
 	RecWin     int    `json:"rec_win" flag:"rec_win" default:"1024" usage:""`
-	Interval   int    `json:"interval" flag:"interval" default:"100" usage:""`
+	Interval   int64    `json:"interval" flag:"interval" default:"100" usage:""`
 	Debug      bool   `json:"debug" flag:"debug" default:"false" usage:""`
 	PPort      int    `json:"p_port" flag:"p_port" default:"8888" usage:""`
 }
