+++ server/config/config.go
@@ -12,7 +12,7 @@ type config struct {
 	Key        string `json:"key" flag:"key" usage:"" default:"sharpshooter"`
 	SendWin    int    `json:"send_win" flag:"send_win" default:"1024" usage:""`
 	RecWin     int    `json:"rec_win" flag:"rec_win" default:"1024" usage:""`
-	Interval   int    `json:"interval" flag:"interval" default:"100" usage:""`
+	Interval   int64    `json:"interval" flag:"interval" default:"100" usage:""`
 	ListenPort int    `json:"listen_port" flag:"listen_port" default:"0" usage:""`
 	Debug      bool   `json:"debug" flag:"debug" default:"false" usage:""`
 	PPort      int    `json:"p_port" flag:"p_port" default:"8888" usage:""`
